LayoutTests/fast/profiler/multiple-anonymous-functions-called-from-the-same-function-expected.txt

This page's JavaScript calls the different anonymous functions the same scope. 

To run this test manually, load it in the browser then load the WebInspector and look at the profile. In the profile there should be a call two entries named to (anonymous function), as opposed to one entry called twice.

Profile title: Multiple calls to different anonymous functions
Thread_1 (no file) (line 0)
   startTest multiple-anonymous-functions-called-from-the-same-function.html (line 11)
      anonymousFunction profiler-test-JS-resources.js (line 29)
         insertNewText profiler-test-JS-resources.js (line 17)
            createElement (no file) (line 0)
            createTextNode (no file) (line 0)
            appendChild (no file) (line 0)
            getElementById (no file) (line 0)
      anotherAnonymousFunction profiler-test-JS-resources.js (line 30)
         insertGivenText profiler-test-JS-resources.js (line 9)
            createElement (no file) (line 0)
            createTextNode (no file) (line 0)
            appendChild (no file) (line 0)
            getElementById (no file) (line 0)
      endTest profiler-test-JS-resources.js (line 1)
